+++ server/src/models/comment.py
@@ -1,4 +1,6 @@
-from sqlalchemy import Column, Integer, ForeignKey, Text
+import datetime
+
+from sqlalchemy import Column, Integer, ForeignKey, Text, DateTime
 from sqlalchemy.orm import relationship
 
 from src.database import Base
@@ -9,6 +11,10 @@ class Comment(Base):
     __tablename__ = "comments"
 
     id = Column(Integer, primary_key=True, index=True)
+    created_at = Column(DateTime, default=datetime.datetime.utcnow)
+    updated_at = Column(
+        DateTime, default=datetime.datetime.utcnow, onupdate=datetime.datetime.utcnow
+    )
     author_id = Column(Integer, ForeignKey("members.id"), nullable=False)
     action_id = Column(Integer, ForeignKey("actions.id"), nullable=False)
     flight_id = Column(Integer, ForeignKey("flights.id"), nullable=True)